feat: added splash screen to verify if user is logged and it's role
